Removed gradle subprojects in favour of sourceSets
