--- build_orm5.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
- ./gradlew -p orm5 --no-daemon clean runShadow --args='-wi 10 -i 20 -f 1 -o orm5_report.jmh'
