+++ gradle/java-module.gradle
@@ -15,3 +15,8 @@ apply plugin: 'project-report'
 apply from: rootProject.file( 'gradle/libraries.gradle' )
 
 sourceCompatibility = 1.8
+
+task wrapper( type: Wrapper ) {
+	gradleVersion = '4.10.2'
+	distributionType = 'ALL'
+}
