--- hql-benchmark-common/hql-benchmark-common.gradle
@@ -1,9 +1,0 @@
-apply from: rootProject.file( 'gradle/java-module.gradle' )
-
-apply plugin: "me.champeau.gradle.jmh"
-
-dependencies {
-    compile 'org.openjdk.jmh:jmh-core:1.21'
-    compile 'javax.persistence:javax.persistence-api:2.2'
-	compile libraries.logging
-}
