--- hql-benchmark-common/src/main/java/org/hibernate/benchmarks/HqlParseTreeBuilder.java
@@ -1,18 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
- */
-package org.hibernate.benchmarks;
-
-/**
- * Delegate for triggering Antlr to parse the HQL into Antlr's "parse tree"
- *
- * @author Andrea Boriero
- * @author Steve Ebersole
- */
-public interface HqlParseTreeBuilder {
-	Object getHQLParser(String hqlString);
-}
-
