--- hql-benchmark-common/src/main/java/org/hibernate/benchmarks/HqlSemanticInterpreter.java
@@ -1,19 +1,0 @@
-/*
- * Hibernate, Relational Persistence for Idiomatic Java
- *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later
- * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
- */
-package org.hibernate.benchmarks;
-
-/**
- * Delegate for performing the initial semantic analysis of the
- * parse tree generated by Antlr from the HQL
- *
- * @author Andrea Boriero
- * @author Steve Ebersole
- */
-public interface HqlSemanticInterpreter {
-	Object getSemanticModel(String hqlString);
-}
-
