--- hql-benchmark-orm5/hql-benchmark-orm5.gradle
@@ -1,6 +1,0 @@
-apply from: rootProject.file( 'gradle/benchmark-module.gradle' )
-
-dependencies {
-	compile libraries.orm5
-	compile project( ':hql-benchmark-common' )
-}
