--- hql-benchmark-orm5/src/main/java/org/hibernate/benchmarks/ParseTreeBuilderVersion5.java
@@ -1,9 +1,0 @@
-package org.hibernate.benchmarks;
-
-public class ParseTreeBuilderVersion5 implements HqlParseTreeBuilder {
-
-    @Override
-    public Object getHQLParser(String hqlString) {
-        return 0;
-    }
-}
