--- hql-benchmark-orm6/hql-benchmark-orm6.gradle
@@ -1,7 +1,0 @@
-
-apply from: rootProject.file( 'gradle/benchmark-module.gradle' )
-
-dependencies {
-	compile libraries.orm6
-	compile project( ':hql-benchmark-common' )
-}
