+++ orm5_benchmarks.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+./gradlew clean jmh --no-daemon -Porm5
