+++ settings.gradle
@@ -1,9 +1,4 @@
-rootProject.name = 'hql-benchmark'
-
-include 'hql-benchmark-common'
-include 'hql-benchmark-orm5'
-include 'hql-benchmark-orm6'
-
+rootProject.name = 'orm-benchmark'
 
 rootProject.children.each { project ->
 	project.buildFileName = "${project.name}.gradle"
@@ -11,4 +6,3 @@ rootProject.children.each { project ->
 	assert project.buildFile.exists()
 	assert project.buildFile.isFile()
 }
-
