+++ src/main/java/org/hibernate/benchmarks/common/HqlBenchmarkTests.java
@@ -1,0 +1,13 @@
+/*
+ * Hibernate, Relational Persistence for Idiomatic Java
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
+ */
+package org.hibernate.benchmarks.common;
+
+/**
+ * @author Steve Ebersole
+ */
+public class HqlBenchmarkTests {
+}
