+++ src/main/java/org/hibernate/benchmarks/model/Address.java
