+++ src/main/java/org/hibernate/benchmarks/model/Animal.java
