+++ src/main/java/org/hibernate/benchmarks/model/Cat.java
