+++ src/main/java/org/hibernate/benchmarks/model/Dog.java
