+++ src/main/java/org/hibernate/benchmarks/model/DomesticAnimal.java
