+++ src/main/java/org/hibernate/benchmarks/model/Human.java
