+++ src/main/java/org/hibernate/benchmarks/model/Joiner.java
