+++ src/main/java/org/hibernate/benchmarks/model/Lizard.java
