+++ src/main/java/org/hibernate/benchmarks/model/Mammal.java
