+++ src/main/java/org/hibernate/benchmarks/model/Name.java
