+++ src/main/java/org/hibernate/benchmarks/model/Reptile.java
