+++ src/main/java/org/hibernate/benchmarks/model/StateProvince.java
