+++ src/main/java/org/hibernate/benchmarks/model/User.java
