+++ src/orm5/java/org/hibernate/benchmarks/ParseTreeBuilderVersion5.java
@@ -1,0 +1,11 @@
+package org.hibernate.benchmarks;
+
+import org.hibernate.benchmarks.common.HqlParseTreeBuilder;
+
+public class ParseTreeBuilderVersion5 implements HqlParseTreeBuilder {
+
+    @Override
+    public Object getHQLParser(String hqlString) {
+        return 0;
+    }
+}
